search

+++ Codeforces/A_Power_Of_Two.cpp
@@ -5,9 +5,9 @@ using namespace std;
 int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
-    int n;
+    long long int n;
     cin>>n;
-    int ans = log2(n);
+    long long int ans = log2(n);
     if(pow(2,ans)==n)cout<<"YES\n";
     else cout<<"NO\n";
     return 0;
