+++ Codeforces/B_Prime_checking.cpp
@@ -17,7 +17,7 @@ void isPrime(long long int n)
         cout << "NO\n";
         return;
     }
-    for (int i = 3; i * i <= n; i+=2)
+    for (long long i = 3; i * i <= n; i+=2)
     {
         if (n % i == 0)
         {
